// src/types/controllers/ControllerCar.ts
/*type DbCar = {
	name: string;
	userId: number;
	brand: string;
	model: string;
	minBid: number;
	country: string;
	city: string;
	vin: string;
	km: number;
	body: string;
	driveTrain: string;
	transmission: string;
}; */

type DbCar = {
	name: string;
	userId: number;
	brand: string;
	model: string;
	minBid: number;
	country: string;
	city: string;
	vin: string;
	km: number;
	endDate: string;
	body: string;
	driveTrain: string;
	transmission: string;
	exterior: string;
	interior: string;
	highlightsTitle: string;
	equipmentTitle: string;
	serviceHistroy: string;
	ownerShipHistory: string;
};

export default DbCar;